Erased/Changed a lot of comments to be overall cleaner and less confusing. Removed  'def remainingTime()' and broke it up into two methods that return remainingSeconds and Minutes respectively. Hurray for verbosity!
Reworked the way nextSongJob is scheduled (it calls playNextSong() ).
playNextSong() reschedules this job because it has access to the current song's
playTime, and we want the next song to be scheduled at
current_song_playtime + FREQUENCY (in minutes)

this also means that hourlyUpdates doesn't need to worry about this too,
it just calls playNextSong().