Bug Fixes, noticed a few issues while writing up the last commit.
I got smart and realized a lot of the helper methods have no need
to be a part of the class. So that removed a lot of 'self's
and made everything overall cleaner.

Added a neato helper function to convert ints into a better time
formatted string.
e.g. 5 => '05'
So now it's much easier (and cleaner) to create a time like '05:13'
Just do a neataroony timeStr(5) + ':' + timeStr(13)